Adds GameModule abstract class
This adds the "GameModule" abstract class. This class will be subclassed
by custom modules to add functionallity to the game.

For now it only contains one method that takes the game state and
updates it. This will probably change in the future to more fine grained
methods.

I am using a abstract class instead of a role to be able to have default
method implementations. That way modules can ignore to implement some of
the methods.