Uses objects in the API instead of ID:s
The first thought was to use the object ID:s to abstract away the
internals of the GameState class but on second thought it does exactly
the opposite. Now objects is used through the API instead, but the user
must be aware of that they are copies. I must remember to document this.