The game can now be played
By adding a simple text-driven interface the game can now be played.

In this first setup the players should try to find out which citizen is
the double agent. The way that it is accomplished is by picking an agent
from all of the citizens known to the player. This agent will then feed
the player with information on events. All citizens should only
participate in events with citizens of the same allegiance, that way it
should be possible to deduce which agent is the double agent.